Changed to T1+C/T2 input channels, fixed training inefficiency
The model now takes in 3 inputs, T1+C, T2, and the mask. I also
realized that having the inputs be 256x256x256 was a total waste
of compute power because the actual input data from these trials
usually looked something like 512x512x30 or something like that.

Which makes sense since you probably wouldn't take hundreds of MRI
images at a time. Anyways I now have the model take in way less
pixels in the z-dimension and its training a lot better. It's
actually memorizing the training set, but its not generalizing
to the validation or test sets yet